requirements.txt
asgiref==3.6.0
backports.weakref==1.0.post1
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.1
Django==4.1.7
django-cors-headers==3.13.0
django-environ==0.9.0
djangorestframework==3.14.0
drf-yasg==1.21.5
firebase-admin==6.1.0
gcloud==0.17.0
google-api-core==2.11.0
google-api-python-client==2.78.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jws==0.1.3
MarkupSafe==2.1.2
msgpack==1.0.4
oauth2client==3.0.0
packaging==23.0
proto-plus==1.22.2
protobuf==4.22.0rc3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
Pyrebase==3.0.27
python-jwt==2.0.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.7.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
